v1.7 bug fix artillery shells not flying, bug fix machine cannon with upgrade dose no damage, bug fix upgrade buttons cant find info display; bug fix info dosent display correct prize after upgrade lvl 1, removed some unnessesery models, reorderd scripts

+++ Assets/Scripts/Enemy/EnemyController.cs
@@ -75,7 +75,7 @@ public class EnemyController : MonoBehaviour
         if (other.tag == "Shell")
         {
             Destroy(other.gameObject);
-            int sDamage = int.Parse(other.name);
+            float sDamage = float.Parse(other.name);
             enemyHealth -= sDamage;
         }
         if (other.tag == "Enemy" && transform.position.x > other.transform.position.x)
